+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0]
+
 ### Added
 
 - Utility to generate LEM font and palette from image
@@ -71,7 +73,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 First serious release.
 
 
-[Unreleased]: https://github.com/Yamakaky/dcpu/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/Yamakaky/dcpu/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/Yamakaky/dcpu/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/Yamakaky/dcpu/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/Yamakaky/dcpu/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/Yamakaky/dcpu/compare/0.1.0...0.1.1